refactor: split out plugin utilities from manifest
