+++ src/expo/manifest.ts
@@ -1,20 +1,9 @@
-import {
-  resolveConfigPluginFunction,
-  resolveConfigPluginFunctionWithInfo,
-} from '@expo/config-plugins/build/utils/plugin-resolver';
-import { Node, Range } from 'jsonc-parser';
+import { Range } from 'jsonc-parser';
 import { DocumentFilter } from 'vscode';
-
-import { resetModuleFrom } from '../utils/module';
 
 export type FileReference = {
   filePath: string;
   fileRange: Range;
-};
-
-type PluginDefiniton = {
-  nameValue: string;
-  nameRange: Range;
 };
 
 /**
@@ -50,48 +39,4 @@ export function getFileReferences(manifest: string): FileReference[] {
   }
 
   return references;
-}
-
-/**
- * Get the plugin definition from manifest node.
- * Both the `name` and it's `range` include the quotes.
- * This supports different plugin definitions:
- *   - `"plugins": ["./my-plguin.js"]`
- *   - `"plugins": ["expo-camera"]`
- *   - `"plugins": [["expo-camera", [...]]`
- */
-export function getPluginDefinition(plugin: Node): PluginDefiniton {
-  const name = plugin.children?.length ? plugin.children[0] : plugin;
-
-  return {
-    nameValue: name.value,
-    nameRange: {
-      // Exclude the quotes from the range
-      offset: name.offset + 1,
-      length: name.length - 2,
-    },
-  };
-}
-
-/**
- * Try to resolve the config plugin information.
- * This resets previously imported modules to reload this information.
- * When it fails to resolve the config plugin, undefined is returned.
- */
-export function resolvePluginInfo(dir: string, name: string) {
-  resetModuleFrom(dir, name);
-
-  try {
-    return resolveConfigPluginFunctionWithInfo(dir, name);
-  } catch {
-    return undefined;
-  }
-}
-
-/**
- * Try to resolve the actual config plugin function.
- * When it fails to resolve the config plugin, an error is thrown.
- */
-export function resolvePluginFunctionUnsafe(dir: string, name: string) {
-  return resolveConfigPluginFunction(dir, name);
 }
